make context accessible from render command
The code gets a bit pointer-chasey as a result,
but it eliminates a lot of copies from the context
structure (and its children) into the render command,
so there's that.

As a first visible result, we now have access to
the insert mode state from the renderer, so we can
e.g. adjust the cursor shape based on insert mode
-- which is exactly what PC/DOS does!