added Amiga system and fonts
Simulates a pure text screen with 4 bitplanes, mapping the 16
standard colors with intensities similar to what the original
ROMs used. Does *not* implement any fancy AmigaDOS window
decorations (yet?).

Since this is the first system that doesn't have a default font size
(it used 8x8 or 10x9 fonts from day one), a bit of plumbing work
had to be done too -- specifically, the default_screen_size for a
system can now be specified in pixels instead of cells.