added default screens for all systems
Some are pixel-perfect, some are not (Amiga, for example,
is missing the AmigaDOS window decorations and the colored
cursor, and ST is missing the Atari logo). For PC, typical
boot screens of MS-DOS 3.30 and 6.22 (the latter on a
486DX2 system) are used.

To do all of this, a few new markup commands had to be implemented.