another round of blinking refinement
Turns out that on PC, underlining forces lines to be foreground (OR),
but the cursor is XOR'ed against that. To emulate this properly,
the renderer now makes an explicit distinction between underlining
(single scanline) and an "extra line" (any number of scanlines)
for the cursor, with the latter being switchable for OR vs. XOR
semantics.

As a side effect, that makes the code of RF_RenderCell()
even simpler.