added some block editing function to test application
Originally implemented to test the new editing functions,
but since they are useful anyway, why not keep them in
until we have a "proper" replacement with nicer UX?

F2 marks the current cursor position. Use this twice
(at different positions) to select the corners of a
rectangle to manipulate.

F3 copies the rectangle to the current cursor position.

F4 clears the rectangle with the currently selected attributes.